+++ Readme.md
@@ -1,7 +1,7 @@
 # Личный проект «Кексобукинг»
 
 * Студент: [Жанна Халбекова](https://up.htmlacademy.ru/javascript/26/user/1774289).
-* Наставник: Андрей Ефимов. (https://htmlacademy.ru/profile/efiand)
+* Наставник: Андрей Ефимов. (https://up.htmlacademy.ru/profession/frontender/13/htmlcss/33/user/367335)
 
 ---
