+++ js/data.js
@@ -1,0 +1,38 @@
+const LatRang = {
+  MIN: 35.65,
+  MAX: 35.7
+};
+
+const LngRang = {
+  MIN: 139.7,
+  MAX: 139.8
+};
+
+const Price = {
+  MIN: 0,
+  MAX: 1000
+};
+
+const Rooms = {
+  MIN: 1,
+  MAX: 5
+};
+
+const Guests = {
+  MIN: 1000,
+  MAX: 9000
+};
+
+const CHECKIN = ['12:00', '13:00', '14:00'];
+
+const CHECKOUT = ['12:00', '13:00', '14:00'];
+
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg.'
+];
+
+export { LatRang, LngRang, Price, Rooms, Guests, CHECKIN, CHECKOUT, TYPES, FEATURES, PHOTOS };
